requirements.txt
# Discord библиотека
discord.py==2.3.2

# База данных
aiosqlite==0.19.0

# Переменные окружения
python-dotenv==1.0.0

# Дополнительные утилиты
Pillow==10.2.0
aiohttp==3.9.1
